Move more logic into MoveBranch.hs
